* updated the structure of folder models
