+++ src/Exadel.CrazyPrice/Exadel.CrazyPrice.Common/Models/Option/LanguageOption.cs
@@ -1,0 +1,17 @@
+﻿namespace Exadel.CrazyPrice.Common.Models.Option
+{
+    /// <summary>
+    /// Represents the supported languages.
+    /// </summary>
+    public enum LanguageOption
+    {
+        /// <summary>
+        /// English
+        /// </summary>
+        En,
+        /// <summary>
+        /// Russian
+        /// </summary>
+        Ru
+    }
+}
