+++ src/Exadel.CrazyPrice/Exadel.CrazyPrice.Common/Models/Option/SortFieldOption.cs
@@ -1,0 +1,15 @@
+﻿namespace Exadel.CrazyPrice.Common.Models.Option
+{
+    /// <summary>
+    /// Represents the fields to sort.
+    /// </summary>
+    public enum SortFieldOption
+    {
+        NameDiscount,
+        AmountOfDiscount,
+        DateStart,
+        DateEnd,
+        CompanyName,
+        RatingDiscount
+    }
+}
