+++ src/Exadel.CrazyPrice/Exadel.CrazyPrice.Common/Models/Option/SortOption.cs
@@ -1,0 +1,17 @@
+﻿namespace Exadel.CrazyPrice.Common.Models.Option
+{
+    /// <summary>
+    /// Represents the sort order.
+    /// </summary>
+    public enum SortOption
+    {
+        /// <summary>
+        /// Sort in ascending order.
+        /// </summary>
+        Asc,
+        /// <summary>
+        /// Sort in descending order.
+        /// </summary>
+        Desc
+    }
+}
