--- src/Exadel.CrazyPrice/Exadel.CrazyPrice.Common/Models/SortFieldOption.cs
@@ -1,11 +1,0 @@
-﻿namespace Exadel.CrazyPrice.Common.Models
-{
-    public enum SortFieldOption
-    {
-        NameDiscount,
-        DateStart,
-        DateEnd,
-        CompanyName,
-        RatingDiscount
-    }
-}
