--- src/Exadel.CrazyPrice/Exadel.CrazyPrice.Common/Models/SortOption.cs
@@ -1,8 +1,0 @@
-﻿namespace Exadel.CrazyPrice.Common.Models
-{
-    public enum SortOption
-    {
-        Asc,
-        Desc
-    }
-}
